# Change all metric data points to be automatically generated