Before (#660):
```
go test -benchmem -bench=. -run=notests ./internal/data/...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              1000000000               0.256 ns/op           0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                4681989               252 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               4736796               249 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            3651998               331 ns/op             320 B/op          8 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              3555200               336 ns/op             320 B/op          8 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 7.523s
```

After:
```
go test -benchmem -bench=. -run=notests ./internal/data/...                                                           
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              1000000000               0.262 ns/op           0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                4814449               249 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               4815667               253 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            4795528               251 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              4808486               249 ns/op             240 B/op          7 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 6.415s
```